requirements.txt
aiogram==2.25.1
aiohttp==3.8.4
aioschedule==0.5.2
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
Babel==2.9.1
beautifulsoup4==4.12.2
certifi==2022.12.7
charset-normalizer==3.1.0
CurrencyConverter==0.17.7
frozenlist==1.3.3
function==1.2.0
idna==3.4
lxml==4.9.2
magic-filter==1.0.9
multidict==6.0.4
pyTelegramBotAPI==4.11.0
python-decouple==3.8
python-dotenv==1.0.0
pytube==12.1.2
pytz==2023.3
requests==2.30.0
schedule==1.2.0
soupsieve==2.4.1
telebot==0.0.5
urllib3==2.0.2
yarl==1.9.2
